Moving files around. Create a single class SyftSerializable, instead of two separate classes.
