+++ syft/execution/state.py
@@ -1,7 +1,4 @@
 from typing import List
-from typing import Tuple
-from typing import Union
-from typing import Dict
 
 import torch
 
@@ -11,11 +8,10 @@ from syft.workers.abstract import AbstractWorker
 from syft_proto.execution.v1.state_pb2 import State as StatePB
 from syft_proto.execution.v1.state_tensor_pb2 import StateTensor as StateTensorPB
 from syft_proto.types.torch.v1.parameter_pb2 import Parameter as ParameterPB
-from syft.interfaces.protobuf_interface import ProtobufInterface
-from syft.interfaces.msgpack_interface import MsgpackInterface
+from serde.syft_serializable import SyftSerializable
 
 
-class State(ProtobufInterface, MsgpackInterface):
+class State(SyftSerializable):
     """The State is a Plan attribute and is used to send tensors along functions.
 
     It references Plan tensor or parameters attributes using their name, and make
