+++ syft/generic/tensor.py
@@ -1,16 +1,13 @@
-import functools
 from typing import List
 import weakref
 
 import syft as sy
-from syft.generic.object import _apply_args
 from syft.generic.object import AbstractObject
 from syft.generic.object import initialize_object
-from syft.interfaces.protobuf_interface import ProtobufInterface
-from syft.interfaces.msgpack_interface import MsgpackInterface
+from serde.syft_serializable import SyftSerializable
 
 
-class AbstractTensor(AbstractObject, ProtobufInterface, MsgpackInterface):
+class AbstractTensor(AbstractObject, SyftSerializable):
     def __init__(
         self,
         id: int = None,
